refactor(admin): AdminStats sous 300 lignes — éclatement complet (H3)
Décomposition d'AdminStats.vue (582 l) selon le pattern G1/AdminProfile :
- useAdminStats.js : toute la logique <script> (state, loadStats,
  refreshInBackground, refreshData, onMounted, cache + auth).
- StatsMainGrid.vue : les 4 cartes larges (utilisateurs, classes/matières,
  séances/visio, évaluations) + CSS de section déplacé VERBATIM avec son @media.
- StatsSecondaryGrid.vue : les 4 petites cartes (leçons, discussions, heures,
  taux de présence) + CSS de section + @media.
- AdminStats.vue devient orchestrateur (217 l) : header + états (loading/error)
  + info-card année universitaire + câblage des deux grilles.

Parité : mêmes appels (auth.getUser/getMeta, readCache/writeCache 'admin_stats'),
mêmes classes CSS, mêmes textes/formats. CSS déplacé sans retouche.

Dette pré-existante conservée / documentée :
- CSS mort supprimé : .loading-grid (orig. l.345-350) n'avait aucun markup
  cible (l'état de chargement utilise <ContentLoader>, pas .loading-grid).
  Supprimé car scoped sans sélecteur cible. Aucun autre comportement modifié.
- Le fallback de stats reste codé en dur à zéro (pas de vrai endpoint), à
  l'identique de l'original.
